requirements.txt
streamlit
pandas
tweepy
google-generativeai
reportlab
python-dotenv
plotly
altair
fpdf
